Module loki_api::logproto

Structs

Chunk

Delete

DroppedStream

EntryAdapter

GetChunkIDsRequest

GetChunkIDsResponse

LabelPair

LabelRequest

LabelResponse

LegacyLabelPair

LegacyLabelPair exists for backwards compatibility reasons and is deprecated. Do not use.

LegacySample

LegacySample exists for backwards compatibility reasons and is deprecated. Do not use.

PushRequest

PushResponse

QueryRequest

QueryResponse

Sample

SampleQueryRequest

SampleQueryResponse

Series

SeriesIdentifier

SeriesRequest

SeriesResponse

StreamAdapter

TailRequest

TailResponse

TailersCountRequest

TailersCountResponse

TimeSeriesChunk

TransferChunksResponse

Enums

Direction